Bring back react-hot-loader/patch and add feature testing for it
