+++ src/AppContainer.prod.js
@@ -1,0 +1,10 @@
+const React = require('react');
+const { Component } = React;
+
+class AppContainer extends Component {
+  render() {
+    return <this.props.component />;
+  }
+}
+
+module.exports = AppContainer;
